+++ uilocalizer-processor-functionaltest/src/main/java/com/devexperts/uilocalizer/funtest/ButtonConsumer.java
@@ -1,0 +1,19 @@
+/*
+ * #%L
+ * UI Localizer
+ * %%
+ * Copyright (C) 2015 - 2020 Devexperts, LLC
+ * %%
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ * #L%
+ */
+
+package com.devexperts.uilocalizer.funtest;
+
+public class ButtonConsumer {
+    public static String getOkText() {
+        return ButtonsTwo.OKAY;
+    }
+}