+++ uilocalizer-processor-functionaltest/src/main/java/com/devexperts/uilocalizer/funtest/Buttons.java
@@ -1,0 +1,20 @@
+/*
+ * #%L
+ * UI Localizer
+ * %%
+ * Copyright (C) 2015 - 2020 Devexperts, LLC
+ * %%
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ * #L%
+ */
+
+package com.devexperts.uilocalizer.funtest;
+
+import com.devexperts.uilocalizer.Localizable;
+
+public class Buttons {
+    @Localizable("test.util.text.Buttons.Cancel")
+    public static final String CANCEL = "Cancel";
+}